split event_info for 1T/nT
