+++ straxen/plugins/__init__.py
@@ -19,6 +19,9 @@ from . import position_reconstruction
 from . import event_area_per_channel
 from .event_area_per_channel import *
 
+from . import event_info
+from .event_info import *
+
 from . import event_patternfit
 from .event_patternfit import *
