+++ straxen/plugins/event_info.py
@@ -1,0 +1,49 @@
+import strax
+from straxen import pre_apply_function
+export, __all__ = strax.exporter()
+
+
+@strax.takes_config(
+    strax.Option(
+        name='event_info_function',
+        default='pre_apply_function',
+        help="Function that must be applied to all event_info data. Do not change.",
+    )
+)
+class EventInfo(strax.MergeOnlyPlugin):
+    """
+    Plugin which merges the information of all event data_kinds into a
+    single data_type.
+    """
+    depends_on = ['events',
+                  'event_basics', 'event_positions', 'corrected_areas',
+                  'energy_estimates',
+                  'event_pattern_fit',
+                  'veto_proximity',
+                  ]
+    save_when = strax.SaveWhen.ALWAYS
+    __version__ = '0.0.1'
+
+    def compute(self, **kwargs):
+        event_info_function = self.config['event_info_function']
+        event_info = super().compute(**kwargs)
+        if event_info_function != 'disabled':
+            event_info = pre_apply_function(event_info,
+                                            self.run_id,
+                                            self.provides,
+                                            event_info_function,
+                                            )
+        return event_info
+
+
+@export
+class EventInfo1T(strax.MergeOnlyPlugin):
+    """
+    Plugin which merges the information of all event data_kinds into a
+    single data_type.
+    """
+    depends_on = ['events',
+                  'event_basics', 'event_positions', 'corrected_areas',
+                  'energy_estimates']
+    save_when = strax.SaveWhen.ALWAYS
+    __version__ = '0.0.0'
