+++ straxen/plugins/event_processing.py
@@ -572,33 +572,3 @@ class EnergyEstimates(strax.Plugin):
 
     def cs2_to_e(self, x):
         return self.config['lxe_w'] * x / self.config['g2']
-
-
-@strax.takes_config(
-    strax.Option(
-        name='event_info_function',
-        default='pre_apply_function',
-        help="Function that must be applied to all event_info data. Do not change.",
-    )
-)
-class EventInfo(strax.MergeOnlyPlugin):
-    """
-    Plugin which merges the information of all event data_kinds into a
-    single data_type.
-    """
-    depends_on = ['events',
-                  'event_basics', 'event_positions', 'corrected_areas',
-                  'energy_estimates']
-    save_when = strax.SaveWhen.ALWAYS
-    __version__ = '0.0.1'
-
-    def compute(self, **kwargs):
-        event_info_function = self.config['event_info_function']
-        event_info = super().compute(**kwargs)
-        if event_info_function != 'disabled':
-            event_info = pre_apply_function(event_info,
-                                            self.run_id,
-                                            self.provides,
-                                            event_info_function,
-                                            )
-        return event_info
